Add four basic operators, and implement the operate and populate functions.
This commit Implements four basic operator functions (Add Subtract Multiply Divide) that you would find
on any calculator, and also two other new functions populate() and Operate().

Populate is the function which populates the calculator screen and the current value the user is
manipulating (displayValue). In essence populate is responsible for mapping all the button inputs
to their correct outputs, and changing various related values, and whats seen on the screen.

The operate function is a simple function which calls the relevant operator when the equals button is
pressed.